added gemini to SchedulePage
