+++ composeApp/src/commonMain/kotlin/pages/HealthPage.kt
@@ -13,7 +13,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import components.CalendarDataSource
-import components.ScheduleView
+
 
 @Composable
 fun HealthPage(onNavigateToTimerView: () -> Unit) {
